Avoid circular import issue.
